#2517 move some of the api requests away from actions
