+++ src/api/window.js
@@ -1,5 +1,8 @@
 import { post, get } from 'axios';
 
+import { getData } from '../actions/GenericActions';
+import { parseToDisplay } from '../utils/documentListHelper';
+
 export function initLayout(
   entity,
   docType,
@@ -32,4 +35,134 @@ export function topActionsRequest(windowId, documentId, tabId) {
   return get(`
     ${config.API_URL}/window/${windowId}/${documentId}/${tabId}/topActions
   `);
+}
+
+
+export function getZoomIntoWindow(
+  entity,
+  windowId,
+  docId,
+  tabId,
+  rowId,
+  field
+) {
+  return get(
+    config.API_URL +
+      '/' +
+      entity +
+      '/' +
+      windowId +
+      (docId ? '/' + docId : '') +
+      (tabId ? '/' + tabId : '') +
+      (rowId ? '/' + rowId : '') +
+      '/field' +
+      '/' +
+      field +
+      '/zoomInto?showError=true'
+  );
+}
+
+export function discardNewRow({ windowType, documentId, tabId, rowId } = {}) {
+  return post(
+    config.API_URL +
+      '/window/' +
+      windowType +
+      '/' +
+      documentId +
+      '/' +
+      tabId +
+      '/' +
+      rowId +
+      '/discardChanges'
+  );
+}
+
+export function discardNewDocument({ windowType, documentId } = {}) {
+  return post(
+    config.API_URL +
+      '/window/' +
+      windowType +
+      '/' +
+      documentId +
+      '/discardChanges'
+  );
+}
+
+export function getTab(tabId, windowType, docId, orderBy) {
+  return getData(
+    'window',
+    windowType,
+    docId,
+    tabId,
+    null,
+    null,
+    null,
+    null,
+    orderBy
+  ).then(
+    res =>
+      res.data &&
+      res.data.map(row => ({
+        ...row,
+        fieldsByName: parseToDisplay(row.fieldsByName),
+      }))
+  );
+}
+
+export function startProcess(processType, pinstanceId) {
+  return get(`${config.API_URL}/process/${processType}/${pinstanceId}/start`);
+}
+
+export function getProcessData({
+  processId,
+  viewId,
+  type,
+  ids,
+  tabId,
+  rowId,
+  selectedTab,
+  childViewId,
+  childViewSelectedIds,
+  parentViewId,
+  parentViewSelectedIds,
+}) {
+  const payload = {
+    processId: processId,
+  };
+
+  if (viewId) {
+    payload.viewId = viewId;
+    payload.viewDocumentIds = ids;
+
+    if (childViewId) {
+      payload.childViewId = childViewId;
+      payload.childViewSelectedIds = childViewSelectedIds;
+    }
+
+    if (parentViewId) {
+      payload.parentViewId = parentViewId;
+      payload.parentViewSelectedIds =
+        parentViewSelectedIds instanceof Array
+          ? parentViewSelectedIds
+          : [parentViewSelectedIds];
+    }
+  } else {
+    payload.documentId = Array.isArray(ids) ? ids[0] : ids;
+    payload.documentType = type;
+    payload.tabId = tabId;
+    payload.rowId = rowId;
+  }
+
+  if (selectedTab) {
+    const { tabId, rowIds } = selectedTab;
+
+    if (tabId && rowIds) {
+      payload.selectedTab = {
+        tabId,
+        rowIds,
+      };
+    }
+  }
+
+  return post(`${config.API_URL}/process/${processId}`, payload);
 }
